requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.0
colorclass==2.2.2
docopt==0.6.2
idna==3.3
influxdb-client==1.30.0
msgpack==1.0.4
packaging==21.3
pip-upgrader==1.4.15
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.1
Rx==3.2.0
six==1.16.0
terminaltables==3.1.10
urllib3==1.26.10
